feat: 本番環境用のDocker設定と動作確認手順書の追加
- 新しいdocker-compose.production.ymlファイルを作成し、フロントエンド、バックエンド、データベース、キャッシュ、検索エンジン、リバースプロキシのサービスを定義
- 最終動作確認手順書「Final_Verification_Guide.md」を追加し、システムの動作確認手順を詳細に文書化
- 本番デプロイチェックリスト「Production_Deployment_Checklist.md」を作成し、デプロイ前の確認項目を整理

これにより、本番環境へのデプロイが円滑に行えるようになり、動作確認の手順が明確化されます。
