+++ _本番リリース前準備/Final_Verification_Guide.md
@@ -1,0 +1,259 @@
+# 最終動作確認手順書
+
+**作成日**: 2025-08-26  
+**対象システム**: SESスキル管理システム  
+**バージョン**: 1.0.0
+
+## 📋 事前準備確認
+
+### 環境変数確認
+```bash
+# backend/.env.production が設定されていることを確認
+cat backend/.env.production | grep USE_NEW_BP_API
+# 期待値: USE_NEW_BP_API=true
+```
+
+### データベース状態確認
+```bash
+# Prismaスキーマが最新であることを確認
+cd backend
+npx prisma validate
+
+# マイグレーションステータス確認
+npx prisma migrate status
+```
+
+## 🧪 機能テスト手順
+
+### 1. 認証機能確認
+
+#### ログイン機能
+1. ブラウザで http://localhost を開く
+2. 以下のアカウントでログイン
+   - Email: admin@example.com
+   - Password: （設定されたパスワード）
+3. ダッシュボードが表示されることを確認
+
+#### セッション管理
+- ブラウザをリロードしてもログイン状態が維持されることを確認
+- ログアウトボタンクリックでログアウトできることを確認
+
+### 2. 取引先企業管理機能
+
+#### 一覧表示
+1. 「取引先管理」メニューをクリック
+2. 取引先企業一覧が表示されることを確認
+3. 以下の項目が表示されていることを確認：
+   - 会社名
+   - 業界
+   - 稼働人数
+   - 月次売上
+   - 評価
+   - ステータス
+
+#### 検索機能
+1. 検索ボックスに文字を入力
+2. リアルタイムで検索結果が絞り込まれることを確認
+3. 業界フィルタを選択して絞り込みができることを確認
+
+#### ページネーション
+1. 11件以上のデータがある場合、ページネーションが表示されることを確認
+2. ページ切り替えが正常に動作することを確認
+
+#### 詳細表示
+1. 一覧から任意の取引先をクリック
+2. 詳細情報が表示されることを確認：
+   - 基本情報（会社名、住所、連絡先等）
+   - 担当者情報
+   - 取引情報（月次売上、稼働人数等）
+   - メモ・備考
+
+#### 新規作成
+1. 「新規登録」ボタンをクリック
+2. 必須項目を入力：
+   - 会社名: テスト株式会社
+   - 会社名カナ: テストカブシキガイシャ
+   - 業界: IT
+3. 「登録」ボタンをクリック
+4. 成功メッセージが表示されることを確認
+5. 一覧に新規作成した企業が表示されることを確認
+
+#### 編集機能
+1. 一覧から任意の取引先の編集ボタンをクリック
+2. 情報を変更（例: 月次売上を変更）
+3. 「更新」ボタンをクリック
+4. 成功メッセージが表示されることを確認
+5. 変更が反映されていることを確認
+
+#### 削除機能
+1. 一覧から任意の取引先の削除ボタンをクリック
+2. 確認ダイアログが表示されることを確認
+3. 「削除」をクリック
+4. 成功メッセージが表示されることを確認
+5. 一覧から削除されていることを確認
+
+### 3. エンジニア管理機能
+
+#### エンジニア一覧
+1. 「エンジニア管理」メニューをクリック
+2. エンジニア一覧が表示されることを確認
+3. 稼働状況（稼働中/待機中）が表示されることを確認
+
+#### エンジニア検索
+1. 検索ボックスでエンジニアを検索
+2. スキルでの絞り込みができることを確認
+3. 稼働状況でのフィルタリングができることを確認
+
+### 4. スキルシート機能
+
+#### スキルシート一覧
+1. 「スキルシート」メニューをクリック
+2. スキルシート一覧が表示されることを確認
+
+#### スキルシート作成・編集
+1. 「新規作成」ボタンをクリック
+2. 基本情報、スキル、プロジェクト履歴を入力
+3. 保存後、PDFエクスポート機能が動作することを確認
+
+## 🔍 APIレスポンステスト
+
+### curlでのAPI確認
+```bash
+# 認証トークン取得
+TOKEN=$(curl -s -X POST http://localhost:8000/api/v1/auth/login \
+  -H "Content-Type: application/json" \
+  -d '{"email":"admin@example.com","password":"password"}' \
+  | jq -r '.token')
+
+# 取引先一覧取得
+curl -H "Authorization: Bearer $TOKEN" \
+  http://localhost:8000/api/v1/business-partners | jq
+
+# 期待されるレスポンス構造:
+# {
+#   "data": [
+#     {
+#       "id": "123",
+#       "companyName": "会社名",
+#       "industry": "IT",
+#       "currentEngineers": 5,
+#       "monthlyRevenue": 1000000,
+#       ...
+#     }
+#   ],
+#   "total": 10
+# }
+```
+
+## 📊 パフォーマンステスト
+
+### レスポンス時間測定
+```bash
+# 取引先一覧API（100件取得）
+time curl -H "Authorization: Bearer $TOKEN" \
+  "http://localhost:8000/api/v1/business-partners?limit=100"
+# 期待値: 1秒以内
+
+# エンジニア一覧API（50件取得）
+time curl -H "Authorization: Bearer $TOKEN" \
+  "http://localhost:8000/api/v1/engineers?limit=50"
+# 期待値: 1秒以内
+```
+
+### データベースクエリ確認
+```bash
+# Prismaクエリログを有効化して確認
+export DATABASE_URL="postgresql://..."
+export DEBUG="prisma:query"
+npm run dev
+
+# ログでN+1問題が発生していないことを確認
+```
+
+## 🔐 セキュリティ確認
+
+### 認証・認可
+- [ ] 未認証でAPIアクセスが拒否されることを確認
+- [ ] 他社のデータにアクセスできないことを確認
+- [ ] 権限のない操作が拒否されることを確認
+
+### 入力値検証
+- [ ] SQLインジェクション対策の確認（特殊文字入力）
+- [ ] XSS対策の確認（スクリプトタグ入力）
+- [ ] 不正なデータ型の入力が拒否されることを確認
+
+## 🚨 エラーハンドリング確認
+
+### ネットワークエラー
+1. バックエンドサーバーを停止
+2. フロントエンドでの操作を実行
+3. 適切なエラーメッセージが表示されることを確認
+
+### バリデーションエラー
+1. 必須項目を空にして送信
+2. 適切なバリデーションメッセージが表示されることを確認
+
+## ✅ 最終チェックリスト
+
+### 機能面
+- [ ] すべての主要機能が正常動作する
+- [ ] エラーハンドリングが適切に実装されている
+- [ ] ユーザビリティに問題がない
+
+### パフォーマンス
+- [ ] ページ読み込み時間が3秒以内
+- [ ] API応答時間が1秒以内
+- [ ] 大量データ処理時も応答性が保たれる
+
+### セキュリティ
+- [ ] 認証・認可が正しく機能している
+- [ ] 入力値検証が実装されている
+- [ ] 機密情報が露出していない
+
+### データ整合性
+- [ ] データマイグレーションが完了している
+- [ ] 既存データが正しく表示される
+- [ ] 新規データの作成・更新・削除が正常動作する
+
+### 互換性
+- [ ] Chrome最新版で動作確認
+- [ ] Firefox最新版で動作確認
+- [ ] Safari最新版で動作確認
+- [ ] Edge最新版で動作確認
+
+## 📝 確認実施記録
+
+| 項目 | 実施日時 | 実施者 | 結果 | 備考 |
+|------|---------|--------|------|------|
+| 認証機能 |  |  | ⬜ |  |
+| 取引先管理 |  |  | ⬜ |  |
+| エンジニア管理 |  |  | ⬜ |  |
+| スキルシート |  |  | ⬜ |  |
+| APIテスト |  |  | ⬜ |  |
+| パフォーマンス |  |  | ⬜ |  |
+| セキュリティ |  |  | ⬜ |  |
+| エラー処理 |  |  | ⬜ |  |
+
+## 🔧 トラブルシューティング
+
+### よくある問題と解決方法
+
+#### データが表示されない
+1. コンソールログを確認
+2. ネットワークタブでAPIレスポンスを確認
+3. データベースに実際にデータが存在するか確認
+
+#### ログインできない
+1. 環境変数JWT_SECRETが設定されているか確認
+2. パスワードハッシュが正しいか確認
+3. ユーザーのis_activeがtrueになっているか確認
+
+#### パフォーマンスが遅い
+1. データベースインデックスが適用されているか確認
+2. N+1問題が発生していないか確認
+3. キャッシュが有効になっているか確認
+
+---
+
+**確認完了署名**：_______________  
+**確認日時**：_______________
