+++ _本番リリース前準備/Production_Deployment_Checklist.md
@@ -1,0 +1,158 @@
+# 本番デプロイチェックリスト
+
+**作成日**: 2025-08-26  
+**対象システム**: SESスキル管理システム  
+**バージョン**: 1.0.0
+
+## 🚀 デプロイ前チェックリスト
+
+### 1. 環境設定
+- [ ] `.env.production`ファイルの設定値を確認
+  - [ ] `DATABASE_URL`を本番DBに変更
+  - [ ] `JWT_SECRET`を強力なランダム文字列に変更
+  - [ ] `CORS_ORIGIN`を本番ドメインに変更
+  - [ ] `FRONTEND_URL`を本番URLに変更
+  - [ ] SMTPサーバー設定を本番用に更新
+- [ ] `USE_NEW_BP_API=true`が設定されていることを確認
+- [ ] `NODE_ENV=production`が設定されていることを確認
+
+### 2. データベース準備
+- [ ] 本番データベースのバックアップを取得
+- [ ] Prismaマイグレーションを実行
+  ```bash
+  npx prisma migrate deploy
+  ```
+- [ ] データマイグレーションスクリプトを実行
+  ```bash
+  npx ts-node scripts/migrate-business-partners.ts migrate
+  ```
+- [ ] データ整合性を確認
+
+### 3. セキュリティチェック
+- [ ] JWT秘密鍵が十分に強力か確認
+- [ ] 環境変数に機密情報が直接記載されていないか確認
+- [ ] CORS設定が適切か確認
+- [ ] 認証・認可が正しく機能しているか確認
+- [ ] SQLインジェクション対策が実装されているか確認
+- [ ] XSS対策が実装されているか確認
+
+### 4. コードベース確認
+- [x] 暫定実装コードが削除されているか確認
+  - [x] `partnerListService.ts`削除済み
+  - [x] `partnerListController.ts`削除済み
+  - [x] `partnerList.routes.ts`削除済み
+- [x] テスト専用エンドポイントが無効化されているか確認
+  - [x] `/api/v1/test/*`ルート無効化済み
+- [x] デバッグ用`console.log`が削除されているか確認
+- [x] エラーハンドリングが適切に実装されているか確認
+
+### 5. パフォーマンス最適化
+- [ ] データベースインデックスが設定されているか確認
+- [ ] N+1問題が解決されているか確認
+- [ ] 不要なログ出力が無効化されているか確認
+- [ ] キャッシュ設定が有効か確認（`ENABLE_CACHING=true`）
+
+### 6. Docker設定
+- [ ] 本番用Dockerfileが最適化されているか確認
+- [ ] マルチステージビルドが使用されているか確認
+- [ ] 不要なパッケージが含まれていないか確認
+- [ ] ヘルスチェックが設定されているか確認
+
+### 7. 監視・ログ設定
+- [ ] ログレベルが`info`に設定されているか確認
+- [ ] エラー監視ツールが設定されているか確認
+- [ ] パフォーマンス監視が設定されているか確認
+- [ ] アラート設定が適切か確認
+
+### 8. バックアップ・リカバリ計画
+- [ ] データベースバックアップ戦略が確立されているか
+- [ ] ロールバック手順が文書化されているか
+- [ ] 災害復旧計画が策定されているか
+
+## 📋 デプロイ手順
+
+### ステップ1: 事前準備
+```bash
+# 1. 最新コードを取得
+git pull origin main
+
+# 2. 依存関係を更新
+npm ci
+
+# 3. ビルド実行
+npm run build
+
+# 4. テスト実行
+npm test
+```
+
+### ステップ2: データベース更新
+```bash
+# 1. バックアップ取得
+pg_dump production_db > backup_$(date +%Y%m%d_%H%M%S).sql
+
+# 2. マイグレーション実行
+npx prisma migrate deploy
+
+# 3. データマイグレーション
+NODE_ENV=production npx ts-node scripts/migrate-business-partners.ts migrate
+```
+
+### ステップ3: アプリケーションデプロイ
+```bash
+# 1. Dockerイメージビルド
+docker build -t skillsheet-app:latest .
+
+# 2. コンテナ起動
+docker-compose -f docker-compose.production.yml up -d
+
+# 3. ヘルスチェック
+curl https://your-domain.com/api/health
+```
+
+### ステップ4: 動作確認
+- [ ] ヘルスチェックエンドポイントが正常応答するか
+- [ ] ログイン機能が動作するか
+- [ ] 取引先一覧が表示されるか
+- [ ] データの作成・更新・削除が動作するか
+- [ ] エラーが適切にハンドリングされるか
+
+## 🔄 ロールバック手順
+
+問題が発生した場合の緊急対応手順：
+
+```bash
+# 1. 前バージョンに戻す
+docker-compose down
+git checkout previous-version-tag
+docker build -t skillsheet-app:rollback .
+docker-compose up -d
+
+# 2. データベースをロールバック（必要な場合）
+psql production_db < backup_YYYYMMDD_HHMMSS.sql
+```
+
+## ⚠️ 重要な注意事項
+
+1. **本番環境の`.env`ファイルは絶対にGitにコミットしない**
+2. **デプロイは必ずメンテナンス時間帯に実施**
+3. **デプロイ前後でパフォーマンスメトリクスを比較**
+4. **ユーザーへの事前通知を忘れずに**
+
+## 📞 緊急連絡先
+
+- システム管理者: [連絡先を記入]
+- データベース管理者: [連絡先を記入]
+- インフラ担当: [連絡先を記入]
+
+## 📝 デプロイログ
+
+| 日時 | バージョン | 担当者 | 状態 | 備考 |
+|------|-----------|--------|------|------|
+| 2025-08-26 | 1.0.0 | - | 準備完了 | 初回本番デプロイ |
+
+---
+
+**最終確認者**: _______________  
+**承認者**: _______________  
+**デプロイ実施日時**: _______________
