+++ backend/.env.production
@@ -1,72 +1,37 @@
-# Production Environment Variables
-# このファイルは本番環境でのみ使用されます
-# 重要: このファイルをGitにコミットしないでください
-
 NODE_ENV=production
 PORT=8000
 
 # Database
-DATABASE_URL=postgresql://skillsheet:YOUR_PROD_DB_PASSWORD@YOUR_PROD_DB_HOST:5432/skillsheet_prod
+DATABASE_URL=postgresql://skillsheet:password@postgres:5432/skillsheet_prod
 
 # Redis
-REDIS_URL=redis://YOUR_PROD_REDIS_HOST:6379
+REDIS_URL=redis://redis:6379
 
-# JWT - 本番環境用の強力なシークレット
-# 重要: 以下の値は必ず変更してください
-JWT_SECRET=wzgJOX9FY89EjRvydfpamEjA7KYaXJhD7EQJw6SGVQhqevVkX02QGaa05xZ1Lj0ByWYyv6cuWVTHQjVeeUti0g==
-JWT_REFRESH_SECRET=wPxcWqDScT6lDDr6vNjbUSbopgYanvyp/e7Yz1PRlDhzfZYOMdojajMoq1BFp3q0n0LTTHN86k637DTZilXnpw==
-JWT_EXPIRES_IN=8h
-REFRESH_TOKEN_EXPIRES_IN=30d
+# JWT（本番環境では強力な秘密鍵を使用すること）
+JWT_SECRET=CHANGE_THIS_TO_STRONG_SECRET_IN_PRODUCTION
+JWT_EXPIRES_IN=7d
 
 # Elasticsearch
-ELASTICSEARCH_URL=http://YOUR_PROD_ES_HOST:9200
+ELASTICSEARCH_URL=http://elasticsearch:9200
 
-# CORS - 本番環境のドメインを設定
+# CORS（本番環境のドメインに変更すること）
 CORS_ORIGIN=https://your-production-domain.com
 
 # Logging
 LOG_LEVEL=info
-LOG_FILE_PATH=/var/log/skillsheet/app.log
 
-# Rate Limiting
-RATE_LIMIT_WINDOW_MS=900000  # 15 minutes
-RATE_LIMIT_MAX_REQUESTS=100  # Max requests per window
-
-# Security Headers
-HELMET_ENABLED=true
-CSRF_ENABLED=true
-
-# Session
-SESSION_SECRET=YOUR_SESSION_SECRET_HERE
-SESSION_MAX_AGE=86400000  # 24 hours
-
-# Email Service (if needed)
-SMTP_HOST=smtp.your-email-provider.com
+# メール送信設定（本番環境用に設定すること）
+SMTP_HOST=smtp.gmail.com
 SMTP_PORT=587
-SMTP_USER=noreply@your-domain.com
-SMTP_PASS=YOUR_SMTP_PASSWORD
+SMTP_USER=your-email@gmail.com
+SMTP_PASSWORD=your-app-password
+SMTP_FROM="SES企業管理システム <noreply@example.com>"
 
-# Monitoring
-SENTRY_DSN=YOUR_SENTRY_DSN_HERE
-NEW_RELIC_LICENSE_KEY=YOUR_NEW_RELIC_KEY_HERE
+# フロントエンドURL（本番環境のURLに変更すること）
+FRONTEND_URL=https://your-production-domain.com
 
-# Feature Flags
-ENABLE_AUDIT_LOG=true
-ENABLE_2FA=false
-ENABLE_API_VERSIONING=true
-
-# API Keys (if needed for external services)
-EXTERNAL_API_KEY=YOUR_API_KEY_HERE
-
-# File Upload
-MAX_FILE_SIZE=10485760  # 10MB
-ALLOWED_FILE_TYPES=image/jpeg,image/png,application/pdf
-
-# Backup
-BACKUP_ENABLED=true
-BACKUP_SCHEDULE=0 2 * * *  # Daily at 2 AM
-
-# Notes:
-# 1. 本番環境にデプロイする前に、必ず全ての値を適切に設定してください
-# 2. このファイルは.gitignoreに追加されていることを確認してください
-# 3. 環境変数は環境に応じて管理サービス（AWS Secrets Manager等）を使用することを推奨します
+# Feature Flags（本番環境設定）
+USE_NEW_BP_API=true  # 新実装を使用
+ENABLE_DETAILED_LOGGING=false
+ENABLE_CACHING=true
+DATA_MIGRATION_MODE=false
